feat: ensure selected option scrolls into view #1340
